+++ src/components/Post/elements/Header.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const Header = styled.div`
+    font-size: 19px;
+    line-height: 25px;
+    font-weight: 400;
+    color: #FFFFFF;
+    position: relative;
+`;
+
+export default Header;
