+++ src/components/Post/elements/PostContent.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const PostContent = styled.div`
+    width: 100%;
+`;
+
+export default PostContent;
